mods to lowpass

+++ lowPass/template.c
@@ -1,4 +1,3 @@
-#include "sf-types.h"
 #include "lowpass-input.h"
 
 int
@@ -6,7 +5,7 @@ main(void)
 {
 	int i;
 	int tailLenght = 1;
-	int output[lowpass_input_len];
+	long output[lowpass_input_len];
 	for (i = 0; i < lowpass_input_len; i++) {
 		for(int j = 0; j < tailLenght; j++){
 			output[i] += lowpass_input[i-j];
